Extract Y Variable from a list of Regression Formulas and Learners

Extract Y Variable from a list of Regression Formulas and Learners

Usage

extract_y_variable(formulas, learner_names, data_colnames, y_variable = NULL)

Arguments

formulas

A vector of formulas used for super learning

learner_names

A character vector of names for the learners

data_colnames

The column names of the dataset for super learning

y_variable

(Optional) the y_variable specified by the user