Softmax

A common transformation used to go from a collection of numbers from R to numbers in [0,1] such that they sum to 1.

Usage

softmax(beta)

Arguments

beta

A vector of numeric values to transform